refactor(features:initializeMode): add/update separated components from `RemoteNodeSection`
