+++ lib/src/features/initialize_mode/presentation/widgets/auth_method_component.dart
@@ -1,0 +1,72 @@
+import 'package:fluent_ui/fluent_ui.dart';
+import 'package:gui/src/core/common/colors/app_colors.dart';
+import 'package:gui/src/core/common/widgets/custom_expandable_widget.dart';
+import 'package:gui/src/core/utils/gen/localization/locale_keys.dart';
+import 'package:pactus_gui_widgetbook/app_styles.dart';
+
+/// ## [AuthMethodComponent] Class Documentation
+///
+/// The `AuthMethodComponent` provides a UI section for selecting
+/// an authentication method.
+///
+/// ### UI Components:
+///
+/// - **Authentication Method Label:**
+///   - Displays the text from `LocaleKeys.auth_method`.
+///   - Styled using `InterTextStyles.captionMedium` with `AppColors.primaryGray`
+///
+/// - **Expandable Authentication Options:**
+///   - Uses `CustomExpandableWidget` to provide a dropdown menu.
+///   - Header displays `LocaleKeys.basic_auth`.
+///   - Body contains authentication method options. example(`Item 1`, `Item 2`)
+///   - Styled with `surface3` color from `SurfacePallet`.
+///
+class AuthMethodComponent extends StatelessWidget {
+  const AuthMethodComponent({
+    super.key,
+  });
+
+  @override
+  Widget build(BuildContext context) {
+    return Row(
+      spacing: 33,
+      crossAxisAlignment: CrossAxisAlignment.start,
+      children: [
+        Text(
+          LocaleKeys.auth_method,
+          style: InterTextStyles.captionMedium.copyWith(
+            color: AppColors.primaryGray,
+          ),
+        ),
+        CustomExpandableWidget(
+          header: LocaleKeys.basic_auth,
+          body: Column(
+            crossAxisAlignment: CrossAxisAlignment.start,
+            children: [
+              Text(
+                'Item 1',
+                style: InterTextStyles.captionMedium.copyWith(
+                  color: AppColors.primaryGray,
+                ),
+              ),
+              Text(
+                'Item 2',
+                style: InterTextStyles.captionMedium.copyWith(
+                  color: AppColors.primaryGray,
+                ),
+              ),
+            ],
+          ),
+          headerColor:
+          AppTheme.of(context).extension<SurfacePallet>()!.surface3!,
+          expandedColor:
+          AppTheme.of(context).extension<SurfacePallet>()!.surface3!,
+          padding: EdgeInsets.all(8),
+          animationDuration: 400,
+          width: 125,
+          maxHeight: 200,
+        ),
+      ],
+    );
+  }
+}
